Creating a Activities class to have less responsabilities on the fitbit class
