+++ src/Api/Fitbit.php
@@ -5,13 +5,6 @@ declare(strict_types=1);
 namespace Namelivia\Fitbit\Api;
 
 use GuzzleHttp\Client;
-use Namelivia\Fitbit\Activity\Activity;
-use Namelivia\Fitbit\Activity\Favorites;
-use Namelivia\Fitbit\Activity\Goals\Goals;
-use Namelivia\Fitbit\Activity\Intraday;
-use Namelivia\Fitbit\Activity\Logs\Logs;
-use Namelivia\Fitbit\Activity\TimeSeries;
-use Namelivia\Fitbit\Activity\Types;
 
 class Fitbit
 {
@@ -19,27 +12,14 @@ class Fitbit
     private $baseUrl = 'https://api.fitbit.com/1/user/';
     private $userId = '-';
     private $client;
-    private $activity;
-    private $timeSeries;
-    private $intraday;
-    private $activityTypes;
-    private $activityLogs;
-    private $favorites;
-    private $goals;
+    private $activities;
 
     public function __construct(Client $client)
     {
         $this->client = $client;
-        $this->activity = new Activity($this);
-        $this->timeSeries = new TimeSeries($this);
-        $this->intraday = new Intraday($this);
-        $this->activityTypes = new Types($this);
-        $this->activityLogs = new Logs($this);
-        $this->favorites = new Favorites($this);
-        $this->goals = new Goals($this);
+        $this->activities = new Activities($this);
     }
 
-    //TODO: Now I want to separate this get, etc.. methods from the Fitbit class
     public function get($url)
     {
         return $this->client->get(
@@ -79,38 +59,8 @@ class Fitbit
         $this->userId = '-';
     }
 
-    public function activity()
+    public function activities()
     {
-        return $this->activity;
-    }
-
-    public function timeSeries()
-    {
-        return $this->timeSeries;
-    }
-
-    public function intraday()
-    {
-        return $this->intraday;
-    }
-
-    public function activityTypes()
-    {
-        return $this->activityTypes;
-    }
-
-    public function activityLogs()
-    {
-        return $this->activityLogs;
-    }
-
-    public function favorites()
-    {
-        return $this->favorites;
-    }
-
-    public function goals()
-    {
-        return $this->goals;
+        return $this->activities;
     }
 }
